Struct riddle::RiddleContext

pub struct RiddleContext<'a> { /* fields omitted */ }

Riddle main thread context. It can be accessed before the main event loop is started via RiddleLib::context and is provided to the event loop update closure.

A context is needed for creating some resources, like Windows.

Implementations

impl<'a> RiddleContext<'a>

pub fn quit(&self)

Issue a quit request. The main loop will terminate once the quit message is processed.

Example

let rdl = RiddleLib::new()?;

rdl.run(move |rdl| {
    // Quit issued
    rdl.quit();
    // The program will continue to execute until the quit event is handled
    // by the main event loop.
})

pub fn event(&self) -> &Event

Get the event associated with this context.

If the context was created by RiddleLib::context the event will be Event::PreRunPlaceholder.

Event::ProcessFrame should be used to execute “per-frame” logic.

pub fn state(&self) -> &RiddleState

The Riddle state, allowing systems to be queried.

pub fn audio(&self) -> &AudioSystem

This is supported on crate feature riddle-audio only.

The audio system.

pub fn input(&self) -> &InputSystem

The input system.

pub fn time(&self) -> &TimeSystem

The time system.

pub fn platform(&self) -> &PlatformSystem

The platform system.