Struct riddle::RiddleLib

pub struct RiddleLib { /* fields omitted */ }

Riddle library initializer

Stores the state of all initialized riddle systems before the main event loop is started with RiddleLib::run or RiddleLib::run_with_err and this object is consumed

Implementations

impl RiddleLib

pub fn new() -> Result<Self, RiddleError>

Initialize Riddle. All available subsystems will be initialized.

pub fn run<F>(self, update: F) -> ! where
F: FnMut(&RiddleContext<'_>) + 'static,

Begin the main event loop of the program.

The update closure is given a RiddleContext which provides access to state which is only available on the main thread, and the event that caused the update closure to be called

Example

// Initialize the library
let rdl = RiddleLib::new()?;

// Start the main event loop
rdl.run(move |rdl| {
    // Application logic
    // [..]
})

pub fn run_with_err<Err: Error, F>(self, update: F) -> ! where
F: FnMut(&RiddleContext<'_>) -> Result<(), Err> + 'static,

Begin the main event loop of the program, with an Error result type, to allow the program to terminate by returning an Error value.

See RiddleLib::run for more details.

Example

// Initialize the library
let rdl = RiddleLib::new()?;

// Start the main event loop
rdl.run_with_err::<RiddleError, _>(move |rdl| {
    // Application logic
    // [..]
    Ok(())
})

pub fn context(&self) -> RiddleContext<'_>

Get a RiddleContext during program initialization, before the main event loop is started.

Example

let rdl =  RiddleLib::new()?;

// Get a context before the application starts the main event loop.
let window: WindowHandle = WindowBuilder::new().build(rdl.context())?;

pub fn state(&self) -> &RiddleState

The system state which is thread safe, and which can be cloned for convenient access throughout the program.

Example

// Initialize the library
let rdl = RiddleLib::new()?;

let rdl_state = rdl.state().clone();
std::thread::spawn(move || {
    // rdl_state is available for accessing input, time, audio, etc.
});